requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
cloudinary==1.26.0
Django==3.2.5
django-bootstrap4==3.0.1
django-cloudinary-storage==0.3.0
django-filter==2.4.0
idna==2.10
Pillow==8.3.1
psycopg2==2.9.1
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
